Get rust_syncode to load with python.
We are using maturin as a build tool here. We'll have to introduce a
pyproject.toml, and generally modify the build process somewhat, but this is
already working.

The main issue is passing objects between Python and Rust. Built-ins like
strings have a default implementation that lets them move back and forth, but
anything we implement ourselves will have to have some more clever way of being
handled.

The step now is to determine where to erect the barrier of abstraction between
the two languages to make this as easy as possible.