Implement dfa_mask_store.
I have no idea if this makes any sense at all, but it does compile. Let me come
up with some test cases. It's also hard-wired to consider all and only accept
sequences of length 2: there's a vestigial argument to set the length of accept
sequences, but I could barely get this to compile. Rust is a crazy language and
my brain is melting. I love it.