Add test case to catch unicode problem.
It turns out that my implementation of dmatch barfs on characters that stretch
across multiple bytes in UTF-8. This is because the DFA matches one byte at a
time; sometimes we end up trying to cut the input in the middle of a byte when
we make the recursive call to dmatch on the unmatched suffix of the input
string. I'll have to think about how to fix this.

On the up-side, I found this bug by trying the grammar mask routine out with a
real vocabulary from a real model, so that I've found this bug is a result and
sign of progress!