Fix unicode problem.
I feel like these comments could be written in a less confrontational tone that
explains what's happening more clearly. The basic problem is that we need to
keep track of both bytes and characters simultaneously, because DFAs munch
bytes, but we need to break the string at character boundaries. This is a big
kludge to make that happen.