Implement caching for DFA construction.
Initial profiling revealed that repeated construction of DFAs consumed >90% of
the time spent computing the store, rendering real workloads
intractable. I add a cache of pre-constructed DFAs, returning one where there is a
hit to the cache and otherwise constructing a new one. The cache begins
empty.

I chose to store the cache as the only member of a DFABuilder struct, only one
of which should struct be instantiated throughout the life of the
program. Rather than passing the builder as an argument to each function
implemented in the library, so that they can use the cache when constructing
DFAs (dmatch, in particular, needs to construct a DFA every time it recurs), I
wrap all the library functions that need access to the cache in a struct.

This choice of how to develop the API, as well as the overall hierarchy of the
structs, are highly debatable: I expect to refine these in the future, in
particular in order to clean up the caller's interface to the library
functionality. Nevertheless, the logic of the program is now correct, and early
benchmarks show promise: on my machine (AMD Ryzen 9 PRO 6950H), constructing a
mask for the JSON grammar (12 terminals) and the deepseek-ai/deepseek-coder-6.7b-instruct
model (32022 tokens in vocab) takes just under 134 seconds.