Comment out broken grammar_mask.
Unfortunately, the basic implementation of reference counting introduced in the
previous commit means that we can't mutate the value pointed to. The solution
might be to implement to DerefMut trait for DFAState.

I'm still not totally happy with the API here, but I'm beginning to accept that
it must be so. It's very Java to have the factory class (DFABuilder) for the
actual class (DFAState). I'll have to think of a more Rustacean way to
implement this abstraction.