Fi hidden weird bug with dmatch, adding test case.
This was shaken out by trying to test cases with a single terminal in the
accept sequence. The issue is that we don't know we've moved past the terminal
we're matching until it's too late; when we slice the string, we want to slice
it before the character that failed to match, not after.