Massively refactor the DFAState struct and fix logical bugs.
It's all kind of one commit here. I refactored the struct to make the logic in
dmatch more legible, which in turn allowed me to shake out some more
bugs. There are a lot of test cases at this point, many of them strictly
redundant. Changing the API is becoming kind of a hassle because of how many
places the logic is called. I don't expect anyone in their right mind to
use the internals of this library in their project, and the scope of the thing
is still relatively limited in scale. The API as presented to the user, with
the weird constructors within constructors, is still ugly.