Remove extraneous convenience methods.
I had these because I forgot to use super::* in the test module, so the methods
weren't available. The semantics of how Rust finds these methods is quite
different from how Java does it: you need to import the relevant trait to use
them at all; they don't come for free with a struct that implements them.

As an API note, when the user wants to call these methods that act on the dfa
within the DFAState struct, they have to reach into the struct's dfa field to
do so. Maybe it'd be cleaner to have the DFAState be an opaque struct, but I
don't mind having the fields public and transparent. This is all API clean up,
though, which can wait till the functionality and performance are correct.