Update json_grammar.lark
Removed choice of empty string values for keys. It does not make sense for keys to have empty string values even though it is allowed by JSON formatting (Thoughts on this?)

Anecdote: The model started giving an unending output like this { "" : { "" : { "" : ... 

Making this change took this choice away from the model and I think once it starts predicting actual non-empty keys, it boosts the probability for it to predict proper values in the next iterations.